We have a working INDEX!!
Following the Django tutorial, we now have a working index. The view
takes the databases model pumps it along with the new index.html
template into the HTML response, and the template when presented has
python code ran within it, with if’s, for’s and all that good stuff.